refactor: 테스트 파일을 로직 테스트와 UI 테스트로 분리
